pyproject.toml
[tool.poetry]
name = "cv-bot"
version = "0.1.0"
description = "This is a streamlit application that allows recruiters to ask questions about my life and my characteristics and get the  answers by a LLM RAG-powered."
authors = ["Leonardo Acquaroli <leonardo.acquaroli00@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.4.0"
langchain-openai = "0.1.7"
python-dotenv = "^1.0.1"
pymongo = "^4.7.2"
chromadb = "^0.5.0"
langchain = "^0.2.1"
langchain-community = "^0.2.1"
streamlit = "^1.35.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
